🐛 bug: 날씨API - Client 로 재조정
위치 탐색의 경우 Client 측에서 되어야함
