+++ src/components/feature/locationWeatherBox.tsx
@@ -5,7 +5,7 @@ import { MapPin } from "lucide-react";
 import ServerLocation from "@/components/feature/serverLocation";
 import WeatherItem from "@/components/feature/weatherApi";
 import { convertLatLngToGrid } from "@/lib/togrid";
-import { getLocationData } from "@/data/functions/weather";
+import { getLocationData } from "@/lib/api/weather";
 
 interface LocationInfo {
   region: string;
@@ -37,7 +37,10 @@ export default function LocationWeatherBox() {
     <>
       <div className="flex items-center gap-1 text-16">
         <MapPin className="w-4 h-4 mr-1" />
-        <ServerLocation city={weatherData?.city ?? "위치 정보 없음"} />
+        <ServerLocation
+          region={weatherData?.region || ""}
+          city={weatherData?.city ?? "위치 정보 없음"}
+        />
       </div>
       <div className="absolute right-4 top-3">
         {coords ? (
